+++ src/connection.js
@@ -1,6 +1,7 @@
 const Environment = require('./env');
 const Utils = require('@openeo/js-commons/src/utils');
 const axios = require('axios').default;
+const StacMigrate = require('@radiantearth/stac-migrate');
 
 const AuthProvider = require('./authprovider');
 const BasicProvider = require('./basicprovider');
@@ -115,17 +116,24 @@ class Connection {
 	/**
 	 * List all collections available on the back-end.
 	 * 
+	 * The collections returned always comply to the latest STAC version (currently 1.0.0-rc.1). 
+	 * 
 	 * @async
 	 * @returns {Promise<Collections>} A response compatible to the API specification.
 	 * @throws {Error}
 	 */
 	async listCollections() {
 		let response = await this._get('/collections');
+		if (Utils.isObject(response.data) && Array.isArray(response.data.collections)) {
+			response.data.collections = response.data.collections.map(collection => StacMigrate.collection(collection));
+		}
 		return response.data;
 	}
 
 	/**
 	 * Get further information about a single collection.
+	 * 
+	 * The collection returned always complies to the latest STAC version (currently 1.0.0-rc.1). 
 	 * 
 	 * @async
 	 * @param {string} collectionId - Collection ID to request further metadata for.
@@ -134,12 +142,14 @@ class Connection {
 	 */
 	async describeCollection(collectionId) {
 		let response = await this._get('/collections/' + collectionId);
-		return response.data;
+		return StacMigrate.collection(response.data);
 	}
 
 	/**
 	 * Loads items for a specific image collection.
 	 * May not be available for all collections.
+	 * 
+	 * The items returned always comply to the latest STAC version (currently 1.0.0-rc.1). 
 	 * 
 	 * This is an experimental API and is subject to change.
 	 * 
@@ -186,6 +196,9 @@ class Connection {
 			}
 
 			let response = await this._get(nextUrl, params);
+			if (Utils.isObject(response.data) && Array.isArray(response.data.features)) {
+				response.data.features = response.data.features.map(item => StacMigrate.item(item));
+			}
 			yield response.data;
 
 			page++;